+++ src/main/java/com/ivan4usa/fp/controllers/RecordController.java
@@ -2,6 +2,7 @@ package com.ivan4usa.fp.controllers;
 
 import com.ivan4usa.fp.entity.Record;
 import com.ivan4usa.fp.services.RecordService;
+import com.ivan4usa.fp.services.UserService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,42 +13,56 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 
+@CrossOrigin
 @RestController
 @RequestMapping("/api/record")
 public class RecordController {
 
     private final Logger logger = LogManager.getLogger(this.getClass());
-
     private final RecordService service;
+    private final UserService userService;
 
     @Autowired
-    public RecordController(RecordService service) {
+    public RecordController(RecordService service, UserService userService) {
         this.service = service;
+        this.userService = userService;
     }
 
     @PostMapping("/all")
     public ResponseEntity<List<Record>> findAll(@RequestBody Long userId) {
+        Long checkUserId = this.userService.getUserId();
+        if (!Objects.equals(checkUserId, userId)) {
+            logger.error("Record is not match to user id of " + userId);
+            return new ResponseEntity("Record is not match to user id", HttpStatus.NOT_ACCEPTABLE);
+        }
         return ResponseEntity.ok(service.findAll(userId));
     }
 
     @PostMapping("/id")
     public ResponseEntity<Record> findById(@RequestBody Long id) {
+        Long userId = this.userService.getUserId();
         Record record = null;
         try {
-            record = service.findById(id);
+            record = service.findById(id).get();
         } catch (NoSuchElementException e) {
-            logger.error("record with id = " + id + " not found");
-            return new ResponseEntity("record not found", HttpStatus.NOT_ACCEPTABLE);
+            logger.error("Record with id = " + id + " not found");
+            return new ResponseEntity("Record not found", HttpStatus.NOT_ACCEPTABLE);
+        }
+        if (!Objects.equals(record.getUserId(), userId)) {
+            logger.error("Record is not match to user id of " + id);
+            return new ResponseEntity("Record is not match to user id", HttpStatus.NOT_ACCEPTABLE);
         }
         return ResponseEntity.ok(record);
     }
 
     @PostMapping("/add")
     public ResponseEntity<Record> add(@RequestBody Record record) {
-        if (record.getId() != null || record.getId() != 0) {
-            logger.error("record id must be null");
-            return new ResponseEntity("record id must be null", HttpStatus.NOT_ACCEPTABLE);
+        Long userId = this.userService.getUserId();
+        if (record.getId() != null && record.getId() != 0) {
+            logger.error("Record id must be null");
+            return new ResponseEntity("Record id must be null", HttpStatus.NOT_ACCEPTABLE);
         }
         if (record.getRecordDate() == null ||
                 record.getAmount() == null ||
@@ -55,15 +70,20 @@ public class RecordController {
                 record.getUserId() == null) {
             logger.error("Some fields of record are empty");
             return new ResponseEntity("Some fields of record are empty", HttpStatus.NOT_ACCEPTABLE);
+        }
+        if (!Objects.equals(record.getUserId(), userId)) {
+            return new ResponseEntity("Record is not match to user id", HttpStatus.NOT_ACCEPTABLE);
         }
         return ResponseEntity.ok(service.add(record));
     }
 
     @PatchMapping("/update")
     public ResponseEntity<Record> update(@RequestBody Record record) {
-        if (record.getId() == null || record.getId() == 0) {
-            logger.error("record id is null");
-            return new ResponseEntity("record id is null", HttpStatus.NOT_ACCEPTABLE);
+        Long userId = this.userService.getUserId();
+        Long id = record.getId();
+        if (record.getId() == null && record.getId() == 0) {
+            logger.error("Record id is null");
+            return new ResponseEntity("Record id is null", HttpStatus.NOT_ACCEPTABLE);
         }
         if (record.getRecordDate() == null ||
                 record.getAmount() == null ||
@@ -72,19 +92,32 @@ public class RecordController {
             logger.error("Some fields of record are empty");
             return new ResponseEntity("Some fields of record are empty", HttpStatus.NOT_ACCEPTABLE);
         }
+        if (!Objects.equals(record.getUserId(), userId)) {
+            return new ResponseEntity("Record is not match to user id", HttpStatus.NOT_ACCEPTABLE);
+        }
+        if (!service.findById(id).isPresent()) {
+            logger.error("Record id" + id + " is not found");
+            return new ResponseEntity("Record id " + id + " is not found", HttpStatus.NOT_ACCEPTABLE);
+        }
         return ResponseEntity.ok(service.update(record));
     }
 
     @DeleteMapping("delete")
-    public ResponseEntity<Record> delete(@RequestBody Long id) {
+    public ResponseEntity<Record> delete(@RequestBody Record record) {
+        Long userId = this.userService.getUserId();
+        Long id = record.getId();
+
         if (id == null || id == 0) {
             return new ResponseEntity("id missed", HttpStatus.NOT_ACCEPTABLE);
+        }
+        if (!Objects.equals(record.getUserId(), userId)) {
+            return new ResponseEntity("Record is not match to user id", HttpStatus.NOT_ACCEPTABLE);
         }
         try {
             service.delete(id);
         } catch (EmptyResultDataAccessException e) {
-            logger.error("record id "+ id + " not found");
-            return new ResponseEntity("record id " + id + " not found", HttpStatus.NOT_ACCEPTABLE);
+            logger.error("Record id "+ id + " not found");
+            return new ResponseEntity("Record id " + id + " not found", HttpStatus.NOT_ACCEPTABLE);
         }
         return new ResponseEntity(HttpStatus.OK);
     }
