+++ src/main/java/com/ivan4usa/fp/controllers/TemplateController.java
@@ -2,6 +2,7 @@ package com.ivan4usa.fp.controllers;
 
 import com.ivan4usa.fp.entity.Template;
 import com.ivan4usa.fp.services.TemplateService;
+import com.ivan4usa.fp.services.UserService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,57 +13,76 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 
+@CrossOrigin
 @RestController
 @RequestMapping("/api/template")
 public class TemplateController {
 
     private final Logger logger = LogManager.getLogger(this.getClass());
-
     private final TemplateService service;
+    private final UserService userService;
 
     @Autowired
-    public TemplateController(TemplateService service) {
+    public TemplateController(TemplateService service, UserService userService) {
         this.service = service;
+        this.userService = userService;
     }
 
     @PostMapping("/all")
     public ResponseEntity<List<Template>> findAll(@RequestBody Long userId) {
+        Long checkUserId = this.userService.getUserId();
+        if (!Objects.equals(checkUserId, userId)) {
+            logger.error("Template is not match to user id of " + userId);
+            return new ResponseEntity("Template is not match to user id", HttpStatus.NOT_ACCEPTABLE);
+        }
         return ResponseEntity.ok(service.findAll(userId));
     }
 
     @PostMapping("/id")
     public ResponseEntity<Template> findById(@RequestBody Long id) {
+        Long userId = this.userService.getUserId();
         Template template = null;
         try {
-            template = service.findById(id);
+            template = service.findById(id).get();
         } catch (NoSuchElementException e) {
-            logger.error("template with id = " + id + " not found");
-            return new ResponseEntity("template not found", HttpStatus.NOT_ACCEPTABLE);
+            logger.error("Template with id = " + id + " not found");
+            return new ResponseEntity("Template not found", HttpStatus.NOT_ACCEPTABLE);
+        }
+        if (!Objects.equals(template.getUserId(), userId)) {
+            logger.error("Template is not match to user id of " + id);
+            return new ResponseEntity("Template is not match to user id", HttpStatus.NOT_ACCEPTABLE);
         }
         return ResponseEntity.ok(template);
     }
 
     @PostMapping("/add")
     public ResponseEntity<Template> add(@RequestBody Template template) {
-        if (template.getId() != null || template.getId() != 0) {
-            logger.error("template id must be null");
-            return new ResponseEntity("template id must be null", HttpStatus.NOT_ACCEPTABLE);
+        Long userId = this.userService.getUserId();
+        if (template.getId() != null && template.getId() != 0) {
+            logger.error("Template id must be null");
+            return new ResponseEntity("Template id must be null", HttpStatus.NOT_ACCEPTABLE);
         }
         if (template.getName() == null ||
                 template.getRecordType() == null ||
                 template.getUserId() == null) {
             logger.error("Some fields of template are empty");
             return new ResponseEntity("Some fields of template are empty", HttpStatus.NOT_ACCEPTABLE);
+        }
+        if (!Objects.equals(template.getUserId(), userId)) {
+            return new ResponseEntity("Template is not match to user id", HttpStatus.NOT_ACCEPTABLE);
         }
         return ResponseEntity.ok(service.add(template));
     }
 
     @PatchMapping("/update")
     public ResponseEntity<Template> update(@RequestBody Template template) {
-        if (template.getId() == null || template.getId() == 0) {
-            logger.error("template id is null");
-            return new ResponseEntity("template id is null", HttpStatus.NOT_ACCEPTABLE);
+        Long userId = this.userService.getUserId();
+        Long id = template.getId();
+        if (template.getId() == null && template.getId() == 0) {
+            logger.error("Template id is null");
+            return new ResponseEntity("Template id is null", HttpStatus.NOT_ACCEPTABLE);
         }
         if (template.getName() == null ||
                 template.getRecordType() == null ||
@@ -70,19 +90,32 @@ public class TemplateController {
             logger.error("Some fields of template are empty");
             return new ResponseEntity("Some fields of template are empty", HttpStatus.NOT_ACCEPTABLE);
         }
+        if (!Objects.equals(template.getUserId(), userId)) {
+            return new ResponseEntity("Template is not match to user id", HttpStatus.NOT_ACCEPTABLE);
+        }
+        if (!service.findById(id).isPresent()) {
+            logger.error("Template id" + id + " is not found");
+            return new ResponseEntity("Template id " + id + " is not found", HttpStatus.NOT_ACCEPTABLE);
+        }
         return ResponseEntity.ok(service.update(template));
     }
 
     @DeleteMapping("delete")
-    public ResponseEntity<Template> delete(@RequestBody Long id) {
+    public ResponseEntity<Template> delete(@RequestBody Template template) {
+        Long userId = this.userService.getUserId();
+        Long id = template.getId();
+
         if (id == null || id == 0) {
-            return new ResponseEntity("id missed", HttpStatus.NOT_ACCEPTABLE);
+            return new ResponseEntity("Template id missed", HttpStatus.NOT_ACCEPTABLE);
+        }
+        if (!Objects.equals(template.getUserId(), userId)) {
+            return new ResponseEntity("Template is not match to user id", HttpStatus.NOT_ACCEPTABLE);
         }
         try {
             service.delete(id);
         } catch (EmptyResultDataAccessException e) {
-            logger.error("template id "+ id + " not found");
-            return new ResponseEntity("template id " + id + " not found", HttpStatus.NOT_ACCEPTABLE);
+            logger.error("Template id "+ id + " not found");
+            return new ResponseEntity("Template id " + id + " not found", HttpStatus.NOT_ACCEPTABLE);
         }
         return new ResponseEntity(HttpStatus.OK);
     }
